msm_fb: display: Don't flush img buffers unless it is cached
There is now a bit in mdp_blit_req.flags to indicate if the input
image buffer is not cached, in which case there's no reason to
flush it for DMA. The upper framework (eg. SurfaceFlinger) should
set this flag accordingly to avoid unnecessary cache flushes.

Also removed the destination cache flush since the composition engine
doesn't read back the framebuffer directly.

Change-Id: I1916c6a51c335eb0bdc0c765319b603deca82f92
Signed-off-by: Michael Burr <mburr@codeaurora.org>