msm: clock: Replace 'auto_off' clock op with 'output_enable/_disable' ops
The 'auto_off' clock op provided the ability to disable a clock
without checking or updating its reference count.  Replace this
with an 'output_enable' and 'output_disable' ops that can be used
to enable/disable a clock without reference count checking.

This functionality is exposed through the clk_output_enable() and
clk_output_disable() APIs. This is needed by the GDFS footswitch
driver which must force clocks on and off, bypassing reference counts.

Change-Id: Ic724127bd01aee4b509ef4311c86a91bb9e23e7c
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>