msm: rotator: Enable/disable rotator power rail as needed
On newer SoC like 8x60, power to the rotator hardware must be turned
off and on explicity.  Introduce use of 'regulator' APIs to turn on
and off power with the clocks.

Previous SoC that automatically turned power on and off as part of
clk_enable() and clk_disable() calls are still supported.  The
rotator driver assumes that if the call to regulator_get() is
unsuccessful, then the rail will be controlled automatically.

Change-Id: I36157f02f8b82f8103d0954c3766000f9e5c5b6d
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>