msm: timer: Fix concurrency bug in sched_clock
The clocksource is perceived as stopped and cached  when the primary
code executing its idle code. The secondary core could read the
hardware clock after the cached value is updated by the primary core
resulting in erroneous time values on subsequent sched_clock.

The fix is to treat current ticks that occur far in the future with
reference to the previous tick as a older tick value.

Change-Id: Ic05719db3f40c38cd436aa4c24f85a237b46e000
Signed-off-by: Maheshkumar Sivasubramanian <msivasub@codeaurora.org>