msm_fb: Enable/disable MDP core power rail as needed
On newer SoC like 8x60, power to the MDP hardware must be turned
off and on explicity.  Introduce use of 'regulator' APIs to turn on
and off power with the clocks. Calls are introduced to both the MDP
and LCD drivers so that MDP is now powered off unless both MDP and
LCDC are idle.

Previous SoC that automatically turned power on and off as part of
clk_enable() and clk_disable() calls are still supported.  The
MDP and LCDC drivers now assume that if the calls to regulator_get()
are unsuccessful, then the rail will be controlled automatically.

Change-Id: Ica87d8ca439130cdd6d2df04801de76a888fbdc9
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>