msm: clock-local: Initialize system vdd cached value to invalid level.
Without this fix, if all clocks enabled during boot are set to frequencies
needing only low system vdd, then the system vdd will not be voted on
during boot.  This could theoretically lead to under volting the clocks or
leaving the system vdd high when it doesn't need to be.

Change-Id: I4406ce4b83f7bb7fb2e74f394666d0dd396b74a2
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>