msm: gpiomux: Collapse config enums into one.
The configuration enumerations represent a single abstraction and
were needlessly split.  The fact that they are handled differently on
different targets is an implementation detail, and to expose this in
the api is needless and inappropriate clutter.  The abstract idea of
what can be done via gpiomux is consistent across all targets and,
as such, should be represented by a single consistent type.

As such, collapse the conflicting abstractions into one consistent one
and move the implementation complexity down into the implementation
(where it belongs) and away from the interface (where it doesn't belong).

Change-Id: I1ea5bbc1847208c9e0b8f508c8e62aee7d29e7db
Signed-off-by: Gregory Bean <gbean@codeaurora.org>