msm: watchdog: Increase watchdog timeout to 670ms
The previous watchdog configuration only allowed for 90ms
of slack time between the watchdog petting timer and the
watchdog timeout. This slack could have easily been
consumed by extra items in the workqueue and scheduling
jitter. The watchdog slack is now increased to 370ms, which
was determined by doubling the minimum timeout needed to
boot to userspace.

Change-Id: I0e9f6bc1e2317a2ce21221103a00eb62914840cf
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>