msm: dma: Fix incorrect tsif crci block sizes
The crci block sizes for the tsif were being misconfigured due to a
change in the register spec for the ADM3. This causes the modem to
misconfigure the block size for the tsif crci. Correct any discrepancy
between our crci table and the modem crci table, and configure the block
sizes ourself rather than relying on the modem to do it correctly.

Change-Id: If5ebec12e993313f7c3339e63e7f6b1a7010d13e
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>